requirements.txt
streamlit>=1.35.0
requests
beautifulsoup4
lxml
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
selenium
webdriver-manager
langgraph